+++ Competitive_Programming/HackerRank/10_days_of_JavaScript/Day_0/hello_world.js
