+++ Competitive_Programming/HackerRank/10_days_of_JavaScript/Day_1/let_and_const.js
