+++ Competitive_Programming/HackerRank/10_days_of_JavaScript/Day_3/throw.js
