+++ Competitive_Programming/HackerRank/10_days_of_JavaScript/Day_4/count_objects.js
