+++ Competitive_Programming/HackerRank/10_days_of_JavaScript/Day_7/regex_3.js
