+++ Competitive_Programming/HackerRank/10_days_of_JavaScript/Day_9/binary_calculator/js/binaryCalculator.js
