+++ Competitive_Programming/HackerRank/minion_game.py
